Move explode tests to own file
